Refactor server runtime for tests
